feeling better about architecture
