+++ controllers/bot.controller.js
@@ -1,10 +1,10 @@
 'use strict';
 
 const tracks = require('../controllers/spotify.controller');
-const SlackAccounts = require('../../models/slackAccounts');
-const SpotifyAccounts = require('../../models/spotifyAccounts');
+const SlackAccounts = require('../models/slackAccounts');
+const SpotifyAccounts = require('../models/spotifyAccounts');
 
-const logger = require('../../../logger');
+const logger = require('../logger');
 
 module.exports = (req, res) => {
   const messageText = req.body.event.text;
@@ -34,3 +34,17 @@ module.exports = (req, res) => {
   // no reason to respond with data to slack at this time so end res 
   res.end();
 };
+
+/*'use strict';
+
+const router = require('express').Router();
+const middleware = require('../middleware');
+
+// API specific middleware
+router.use(middleware.filter);
+router.use(middleware.isAuthorized);
+
+router.use('/bot', require('./bot'));
+
+module.exports = router;
+*/
