+++ routes/middleware/index.js
@@ -30,7 +30,7 @@ exports.handleError = (err, req, res, next) => {
   res.status(500).send({ error: err });
 };
 
-exports.filter = (req, res, next) => {
+exports.slack = (req, res, next) => {
   const isUserGenerated = !!req.body.event && req.body.event.user;
   const isChallengeRequest = req.body.type === 'url_verification';
